+++ DepositApi.DAL/Models/Deposit.cs
@@ -8,12 +8,12 @@ namespace DepositApi.DAL.Models
     {
         public int Id { get; set; }
 
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         public int Term { get; set; }
 
         public DateTime Date { get; set; }
 
-        public double Percent { get; set; }
+        public decimal Percent { get; set; }
     }
 }
